message_logs_app/requirements.txt
Flask==2.0.1
Flask-SocketIO==5.1.1
python-dotenv==0.19.0
psycopg2-binary==2.9.1
aiohttp==3.7.4
Flask-CORS==3.0.10